refactor: presentation サンプルをディレクトリ構造化（既存 7 ファイル移行）
#110 (#111) - presentation サンプル拡張の第一ステップ。

既存 7 サンプルファイルを presentation 値ごとのディレクトリ構造へ
git mv で移行する。挙動・サンプル数は変えず、URL とファイルパスを
最終形（on-root.tsx / in-tab.tsx）に揃えるリファクタ。

- ルート Stack 配下: modal/transparent-modal/full-screen-modal/form-sheet
  を `<value>/on-root.tsx` に移動
- タブ内 Stack 配下: card/contained-modal/contained-transparent-modal
  を `<value>/in-tab.tsx` に移動
- 各 `_layout.tsx` の `<Stack.Screen name>` を新パスに更新
- 一覧画面 (`(tabs)/(home)/navigation-patterns/index.tsx`) は中間ページを
  挟まずサンプル画面へ直結（中間ページは #112 で導入）
- PresentationSampleScreen への相対 import パスを 1 段深く調整
- `lingui:extract` で messages.po のソース参照を新パスに更新
