+++ apps/sandbox/src/app/(tabs)/(home)/navigation-patterns/index.tsx
@@ -6,31 +6,31 @@ export default function NavigationPatterns() {
   const { t } = useLingui();
   const list = [
     {
-      href: "/navigation-patterns/card",
+      href: "/navigation-patterns/card/in-tab",
       text: <Trans>card (push)</Trans>,
     },
     {
-      href: "/navigation-patterns/modal",
+      href: "/navigation-patterns/modal/on-root",
       text: <Trans>modal</Trans>,
     },
     {
-      href: "/navigation-patterns/transparent-modal",
+      href: "/navigation-patterns/transparent-modal/on-root",
       text: <Trans>transparentModal</Trans>,
     },
     {
-      href: "/navigation-patterns/contained-modal",
+      href: "/navigation-patterns/contained-modal/in-tab",
       text: <Trans>containedModal</Trans>,
     },
     {
-      href: "/navigation-patterns/contained-transparent-modal",
+      href: "/navigation-patterns/contained-transparent-modal/in-tab",
       text: <Trans>containedTransparentModal</Trans>,
     },
     {
-      href: "/navigation-patterns/full-screen-modal",
+      href: "/navigation-patterns/full-screen-modal/on-root",
       text: <Trans>fullScreenModal</Trans>,
     },
     {
-      href: "/navigation-patterns/form-sheet",
+      href: "/navigation-patterns/form-sheet/on-root",
       text: <Trans>formSheet</Trans>,
     },
   ] as const satisfies LinkItem[];
